+++ packages/seroval/src/core/UnsupportedTypeError.ts
