+++ packages/seroval/src/core/compat.ts
