+++ packages/seroval/src/core/context.ts
@@ -1,0 +1,4 @@
+export interface BaseParserContext {
+  // Supported features
+  features: number;
+}
