+++ packages/seroval/src/core/reference.ts
