+++ packages/seroval/src/core/string.ts
