+++ packages/seroval/src/core/tree/primitives.ts
@@ -3,170 +3,21 @@ import { Feature } from '../compat';
 import type { ParserContext } from './context';
 import { createIndexedValue } from './context';
 import { serializeString } from '../string';
-import type { BigIntTypedArrayValue, TypedArrayValue } from '../types';
+import type { BigIntTypedArrayValue, TypedArrayValue } from '../../types';
 import UnsupportedTypeError from '../UnsupportedTypeError';
-import {
-  INFINITY_NODE,
-  NEG_INFINITY_NODE,
-  NAN_NODE,
-  NEG_ZERO_NODE,
-} from './constants';
 import { getReferenceID, hasReferenceID } from '../reference';
 import type {
-  SerovalBigIntNode,
   SerovalBigIntTypedArrayNode,
-  SerovalDateNode,
-  SerovalIndexedValueNode,
-  SerovalRegExpNode,
-  SerovalStringNode,
   SerovalTypedArrayNode,
   SerovalWKSymbolNode,
   SerovalReferenceNode,
-  SerovalArrayBufferNode,
   SerovalDataViewNode,
   SerovalNode,
-} from './types';
+  SerovalIndexedValueNode,
+} from '../types';
 import type { WellKnownSymbols } from '../constants';
 import { SerovalNodeType, INV_SYMBOL_REF } from '../constants';
-
-export function createNumberNode(value: number): SerovalNode {
-  switch (value) {
-    case Infinity:
-      return INFINITY_NODE;
-    case -Infinity:
-      return NEG_INFINITY_NODE;
-    default:
-      // eslint-disable-next-line no-self-compare
-      if (value !== value) {
-        return NAN_NODE;
-      }
-      if (Object.is(value, -0)) {
-        return NEG_ZERO_NODE;
-      }
-      return {
-        t: SerovalNodeType.Number,
-        i: undefined,
-        s: value,
-        l: undefined,
-        c: undefined,
-        m: undefined,
-        d: undefined,
-        a: undefined,
-        f: undefined,
-        b: undefined,
-        o: undefined,
-      };
-  }
-}
-
-export function createStringNode(value: string): SerovalStringNode {
-  return {
-    t: SerovalNodeType.String,
-    i: undefined,
-    s: serializeString(value),
-    l: undefined,
-    c: undefined,
-    m: undefined,
-    d: undefined,
-    a: undefined,
-    f: undefined,
-    b: undefined,
-    o: undefined,
-  };
-}
-
-export function createBigIntNode(
-  ctx: ParserContext,
-  current: bigint,
-): SerovalBigIntNode {
-  assert(ctx.features & Feature.BigInt, new UnsupportedTypeError(current));
-  return {
-    t: SerovalNodeType.BigInt,
-    i: undefined,
-    s: '' + current,
-    l: undefined,
-    c: undefined,
-    m: undefined,
-    d: undefined,
-    a: undefined,
-    f: undefined,
-    b: undefined,
-    o: undefined,
-  };
-}
-
-export function createIndexedValueNode(id: number): SerovalIndexedValueNode {
-  return {
-    t: SerovalNodeType.IndexedValue,
-    i: id,
-    s: undefined,
-    l: undefined,
-    c: undefined,
-    m: undefined,
-    d: undefined,
-    a: undefined,
-    f: undefined,
-    b: undefined,
-    o: undefined,
-  };
-}
-
-export function createDateNode(id: number, current: Date): SerovalDateNode {
-  return {
-    t: SerovalNodeType.Date,
-    i: id,
-    s: current.toISOString(),
-    l: undefined,
-    c: undefined,
-    m: undefined,
-    d: undefined,
-    f: undefined,
-    a: undefined,
-    b: undefined,
-    o: undefined,
-  };
-}
-
-export function createRegExpNode(id: number, current: RegExp): SerovalRegExpNode {
-  return {
-    t: SerovalNodeType.RegExp,
-    i: id,
-    s: undefined,
-    l: undefined,
-    c: current.source,
-    m: current.flags,
-    d: undefined,
-    a: undefined,
-    f: undefined,
-    b: undefined,
-    o: undefined,
-  };
-}
-
-export function createArrayBufferNode(
-  id: number,
-  current: ArrayBuffer,
-): SerovalArrayBufferNode {
-  const bytes = new Uint8Array(current);
-  const len = bytes.length;
-  const values = new Array<number>(len);
-  for (let i = 0; i < len; i++) {
-    values[i] = bytes[i];
-  }
-  return {
-    t: SerovalNodeType.ArrayBuffer,
-    i: id,
-    s: values,
-    l: undefined,
-    c: undefined,
-    m: undefined,
-    d: undefined,
-    a: undefined,
-    f: undefined,
-    b: undefined,
-    o: undefined,
-  };
-}
+import { createArrayBufferNode, createIndexedValueNode } from '../base-primitives';
 
 export function serializeArrayBuffer(
   ctx: ParserContext,
@@ -290,7 +141,7 @@ export function createDataViewNode(
 export function createSymbolNode(
   ctx: ParserContext,
   current: symbol,
-): SerovalNode {
+): SerovalIndexedValueNode | SerovalReferenceNode | SerovalWKSymbolNode {
   const id = createIndexedValue(ctx, current);
   if (ctx.markedRefs.has(id)) {
     return createIndexedValueNode(id);
@@ -305,7 +156,7 @@ export function createFunctionNode(
   ctx: ParserContext,
   // eslint-disable-next-line @typescript-eslint/ban-types
   current: Function,
-): SerovalNode {
+): SerovalIndexedValueNode | SerovalReferenceNode {
   assert(hasReferenceID(current), new Error('Cannot serialize function without reference ID.'));
   const id = createIndexedValue(ctx, current);
   if (ctx.markedRefs.has(id)) {
